Extract some of soqlUnpackFeatures
